1. change the H/W and network configurations through the cubemx.   - changed to use HSE clock.   - changed to use the cmsis v1 instead of v2.   - changed configurations in FreeRTOS and lwIP.
