+++ source/application/STM32F439ZI_LwIP_TCPIP/Core/Src/main.cpp
@@ -8,43 +8,18 @@
  ***************************************************************************************************/
 
 /****************************************** Includes ***********************************************/
-#include "main.h"
 #include "common.h"
+#include "main.h"
 #include "stm32f4xx_nucleo_144.h"
 #include "cmsis_os.h"
-#include "lwip.h"
-#include "lwip/tcp.h"
+#include "usart.h"
+#include "gpio.h"
 
-/******************************************** Consts ***********************************************/
-#define ECHO_SERVER_ADDR_0  192
-#define ECHO_SERVER_ADDR_1  168
-#define ECHO_SERVER_ADDR_2  1
-#define ECHO_SERVER_ADDR_3  3
-#define ECHO_SERVER_PORT    7
-
-/**************************************** Static Variables *****************************************/
-UART_HandleTypeDef huart3;
-osThreadId_t defaultTaskHandle;
-const osThreadAttr_t defaultTask_attributes = 
-{
-   .name = "defaultTask",
-   .stack_size = 1024 * 4,
-   .priority = (osPriority_t) osPriorityNormal,
-};
-
-static struct tcp_pcb *echoServerPcb;
-static struct tcp_pcb *clientPcb;
+/**************************************** External Functions ***************************************/
+extern "C" void MX_FREERTOS_Init ( );
 
 /**************************************** Local Functions ******************************************/
-static void    MX_GPIO_Init         ( );
-static void    MX_USART3_UART_Init  ( );
-static void    SystemClock_Config   ( );
-static void    startDefaultTask     ( void *argument );
-
-static void    initTcpEchoServer    ( );
-static err_t   echoAcceptCallback   ( void *arg, struct tcp_pcb *newpcb, err_t err );
-static err_t   echoRecvCallback     ( void *arg, struct tcp_pcb *tpcb, struct pbuf *p, err_t err );
-
+static void SystemClock_Config   ( );
 
 /*************************************** Function Definitions **************************************/
 /**
@@ -56,6 +31,8 @@ int main(void)
    HAL_Init();
    SystemClock_Config();
 
+   MX_FREERTOS_Init();
+
    MX_GPIO_Init();
    MX_USART3_UART_Init();
    BSP_LED_Init( LED_RED );
@@ -64,33 +41,10 @@ int main(void)
 
    LOGGING( "Welcome to STM32F439ZI LwIP TCP/IP Application" );
 
-   osKernelInitialize();
-
-   defaultTaskHandle = osThreadNew( startDefaultTask, NULL, &defaultTask_attributes );
-
    osKernelStart();
 
    while (1)
    { }
-}
-
-/**
- * @brief  Function implementing the defaultTask thread.
- * @param  argument: Not used
- * @retval None
- */
-static void startDefaultTask(void *argument)
-{
-   MX_LWIP_Init();
-
-   initTcpEchoServer();
-
-   for(;;)
-   {    
-      osDelay(1000);
-      LOGGING( "Default Task" );
-      BSP_LED_Toggle( LED_BLUE );
-   }
 }
 
 /**
@@ -110,23 +64,15 @@ static void SystemClock_Config(void)
    /** Initializes the RCC Oscillators according to the specified parameters
     * in the RCC_OscInitTypeDef structure.
     */
-   RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSI;
-   RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-   RCC_OscInitStruct.HSICalibrationValue = RCC_HSICALIBRATION_DEFAULT;
+   RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSE;
+   RCC_OscInitStruct.HSEState = RCC_HSE_ON;
    RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-   RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSI;
-   RCC_OscInitStruct.PLL.PLLM = 8;
-   RCC_OscInitStruct.PLL.PLLN = 180;
+   RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
+   RCC_OscInitStruct.PLL.PLLM = 4;
+   RCC_OscInitStruct.PLL.PLLN = 168;
    RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV2;
-   RCC_OscInitStruct.PLL.PLLQ = 4;
+   RCC_OscInitStruct.PLL.PLLQ = 7;
    if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK)
-   {
-      Error_Handler();
-   }
-
-   /** Activate the Over-Drive mode
-    */
-   if (HAL_PWREx_EnableOverDrive() != HAL_OK)
    {
       Error_Handler();
    }
@@ -147,27 +93,6 @@ static void SystemClock_Config(void)
 }
 
 /**
- * @brief USART2 Initialization Function
- * @param None
- * @retval None
- */
-static void MX_USART3_UART_Init(void)
-{
-   huart3.Instance = USART3;
-   huart3.Init.BaudRate = 115200;
-   huart3.Init.WordLength = UART_WORDLENGTH_8B;
-   huart3.Init.StopBits = UART_STOPBITS_1;
-   huart3.Init.Parity = UART_PARITY_NONE;
-   huart3.Init.Mode = UART_MODE_TX_RX;
-   huart3.Init.HwFlowCtl = UART_HWCONTROL_NONE;
-   huart3.Init.OverSampling = UART_OVERSAMPLING_16;
-   if (HAL_UART_Init(&huart3) != HAL_OK)
-   {
-      Error_Handler();
-   }
-}
-
-/**
  * @brief Redirects the C library printf function to the USART2.
  * @param file: File descriptor (not used)
  * @param ptr: Pointer to the data to be sent
@@ -177,22 +102,6 @@ static void MX_USART3_UART_Init(void)
 extern "C" int _write(int file, char *ptr, int len)
 {
    HAL_UART_Transmit(&huart3, (uint8_t*)ptr, len, HAL_MAX_DELAY);
-}
-
-/**
- * @brief GPIO Initialization Function
- * @param None
- * @retval None
- */
-static void MX_GPIO_Init(void)
-{
-   /* GPIO Ports Clock Enable */
-   __HAL_RCC_GPIOC_CLK_ENABLE();
-   __HAL_RCC_GPIOH_CLK_ENABLE();
-   __HAL_RCC_GPIOA_CLK_ENABLE();
-   __HAL_RCC_GPIOB_CLK_ENABLE();
-   __HAL_RCC_GPIOD_CLK_ENABLE();
-   __HAL_RCC_GPIOG_CLK_ENABLE();
 }
 
 /**
@@ -234,94 +143,3 @@ void assert_failed(uint8_t *file, uint32_t line)
 
 }
 #endif /* USE_FULL_ASSERT */
-
-/**
- * @brief Initializes the TCP Echo Server.
- */
-static void initTcpEchoServer( )
-{
-   ip_addr_t ipAddr;
-   IP4_ADDR( &ipAddr, ECHO_SERVER_ADDR_0, ECHO_SERVER_ADDR_1, ECHO_SERVER_ADDR_2, ECHO_SERVER_ADDR_3 );
-
-   echoServerPcb = tcp_new();
-   if ( echoServerPcb == nullptr ) 
-   {
-      LOGGING( "Failed to create PCB." );
-      return;
-   }
-
-   if ( tcp_bind( echoServerPcb, &ipAddr, ECHO_SERVER_PORT ) == ERR_OK )
-   {
-      echoServerPcb = tcp_listen( echoServerPcb );
-      tcp_accept( echoServerPcb, echoAcceptCallback );
-      LOGGING( "TCPIP: Echo Server is listening on port %d...", ECHO_SERVER_PORT );
-   }
-   else
-   {
-      LOGGING( "TCPIP: Failed to bind PCB." );
-      tcp_abort( echoServerPcb );
-      echoServerPcb = nullptr;
-   }
-}
-
-/**
- * @brief Callback function for accepting new TCP connections.
- * 
- * @param arg argument passed to the callback (not used)
- * @param newpcb pointer to the new TCP PCB
- * @param err error code (not used)
- * @return err_t 
- */
-static err_t echoAcceptCallback( void *arg, struct tcp_pcb *newpcb, err_t err )
-{
-   (void)arg;
-   (void)err;
-
-   LOGGING( "TCPIP: Client connected." );
-   clientPcb = newpcb;
-
-   //!< Set the receive callback for the new PCB
-   tcp_recv( newpcb, echoRecvCallback );
-
-   return ERR_OK;
-}
-
-/**
- * @brief Callback function for receiving data on the TCP connection.
- * 
- * @param arg argument passed to the callback (not used)
- * @param tpcb pointer to the TCP PCB
- * @param p pointer to the received pbuf
- * @param err error code (not used)
- * @return err_t 
- */
-static err_t echoRecvCallback( void *arg, struct tcp_pcb *tpcb, struct pbuf *p, err_t err )
-{
-   if ( err != ERR_OK )
-   {
-      LOGGING( "TCPIP: Receive error: %d", err );
-      if ( p != nullptr )
-      {
-         goto EXIT;
-      }
-      return err;
-   }
-
-   //!< Disconnect if p is null (client disconnected)
-   if ( p == nullptr )
-   {
-      tcp_close( tpcb );
-      LOGGING( "TCPIP: Client disconnected." );
-      return ERR_OK;
-   }
-
-   LOGGING( "TCPIP: Received data: len=%d", p->len );
-
-   //!< Echo the received data back to the client
-   tcp_write( tpcb, p->payload, p->len, 1 );
-   tcp_output( tpcb );
-
-EXIT:   
-   pbuf_free( p );
-   return err;
-}
