+++ source/application/STM32F439ZI_LwIP_TCPIP/STM32F439ZI_LwIP_TCPIP.code-workspace
@@ -1,0 +1,11 @@
+{
+	"folders": [
+		{
+			"path": "../../.."
+		},
+		{
+			"path": "."
+		}
+	],
+   "settings": {}
+}
